// frontend/src/openapi/models/EvaluatorScoreHistory.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Overbae API
 * Backend API for the Overclaw / Overmind platform.
 *
 * The version of the OpenAPI document: 2.0.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { EvaluatorScoreHistoryPoint } from './EvaluatorScoreHistoryPoint';
import {
    EvaluatorScoreHistoryPointFromJSON,
    EvaluatorScoreHistoryPointFromJSONTyped,
    EvaluatorScoreHistoryPointToJSON,
    EvaluatorScoreHistoryPointToJSONTyped,
} from './EvaluatorScoreHistoryPoint';

/**
 *
 * @export
 * @interface EvaluatorScoreHistory
 */
export interface EvaluatorScoreHistory {
    /**
     *
     * @type {string}
     * @memberof EvaluatorScoreHistory
     */
    name: string;
    /**
     *
     * @type {string}
     * @memberof EvaluatorScoreHistory
     */
    evaluatorId: string | null;
    /**
     *
     * @type {Array<EvaluatorScoreHistoryPoint>}
     * @memberof EvaluatorScoreHistory
     */
    points: Array<EvaluatorScoreHistoryPoint>;
}

/**
 * Check if a given object implements the EvaluatorScoreHistory interface.
 */
export function instanceOfEvaluatorScoreHistory(value: object): value is EvaluatorScoreHistory {
    if (!('name' in value) || value['name'] === undefined) return false;
    if (!('evaluatorId' in value) || value['evaluatorId'] === undefined) return false;
    if (!('points' in value) || value['points'] === undefined) return false;
    return true;
}

export function EvaluatorScoreHistoryFromJSON(json: any): EvaluatorScoreHistory {
    return EvaluatorScoreHistoryFromJSONTyped(json, false);
}

export function EvaluatorScoreHistoryFromJSONTyped(json: any, ignoreDiscriminator: boolean): EvaluatorScoreHistory {
    if (json == null) {
        return json;
    }
    return {

        'name': json['name'],
        'evaluatorId': json['evaluator_id'],
        'points': ((json['points'] as Array<any>).map(EvaluatorScoreHistoryPointFromJSON)),
    };
}

export function EvaluatorScoreHistoryToJSON(json: any): EvaluatorScoreHistory {
    return EvaluatorScoreHistoryToJSONTyped(json, false);
}

export function EvaluatorScoreHistoryToJSONTyped(value?: EvaluatorScoreHistory | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {

        'name': value['name'],
        'evaluator_id': value['evaluatorId'],
        'points': ((value['points'] as Array<any>).map(EvaluatorScoreHistoryPointToJSON)),
    };
}